Moved datastructures to seperate directory
