+++ lib/theory/src/ClosedTheory.hs
@@ -3,6 +3,7 @@
 
 module ClosedTheory (
     module ClosedTheory
+    ,module Items.ClosedTheoryItem
 ) where
 
 import           Prelude                             hiding (id, (.))
@@ -25,26 +26,8 @@ import TheoryObject
 import Rule
 import Safe
 
+import Items.ClosedTheoryItem
 
--- | Closed theories can be proven. Invariants:
---     1. Lemma names are unique
---     2. All proof steps with annotated sequents are sound with respect to the
---        closed rule set of the theory.
---     3. Maude is running under the given handle.
-type ClosedTheory =
-    Theory SignatureWithMaude ClosedRuleCache ClosedProtoRule IncrementalProof ()
-
--- | Closed Diff theories can be proven. Invariants:
---     1. Lemma names are unique
---     2. All proof steps with annotated sequents are sound with respect to the
---        closed rule set of the theory.
---     3. Maude is running under the given handle.
-type ClosedDiffTheory =
-    DiffTheory SignatureWithMaude ClosedRuleCache DiffProtoRule ClosedProtoRule IncrementalDiffProof IncrementalProof
-
--- | Either Therories can be Either a normal or a diff theory
-
-type EitherClosedTheory = Either ClosedTheory ClosedDiffTheory
 ------------------------------------------------------------------------------
 -- Closed theory querying / construction / modification
 ------------------------------------------------------------------------------
