+++ lib/theory/src/Items/OpenTheoryItem.hs
@@ -1,0 +1,37 @@
+module Items.OpenTheoryItem (
+    module Items.OpenTheoryItem
+) where
+
+import           Prelude                             hiding (id, (.))
+
+import           Theory.Model
+
+import TheoryObject
+
+import SpecificProof
+
+import Rule
+
+-- | Open theories can be extended. Invariants:
+--   1. Lemma names are unique.
+type OpenTheory =
+    Theory SignaturePure [IntrRuleAC] OpenProtoRule ProofSkeleton SapicElement
+
+type OpenTheoryItem = TheoryItem OpenProtoRule ProofSkeleton SapicElement
+
+
+-- | Open theories can be extended. Invariants:
+--   1. Lemma names are unique.
+--   2. All SapicItems are translated
+type OpenTranslatedTheory =
+    Theory SignaturePure [IntrRuleAC] OpenProtoRule ProofSkeleton ()
+
+-- | Open diff theories can be extended. Invariants:
+--   1. Lemma names are unique.
+type OpenDiffTheory =
+    DiffTheory SignaturePure [IntrRuleAC] DiffProtoRule OpenProtoRule DiffProofSkeleton ProofSkeleton
+
+-- | Either Therories can be Either a normal or a diff theory
+
+-- type EitherTheory = Either Theory  DiffTheory
+type EitherOpenTheory = Either OpenTheory OpenDiffTheory
